+++ lib/get.dart
@@ -5,6 +5,8 @@ export 'src/get_main.dart';
 export 'src/snackbar/snack.dart';
 export 'src/bottomsheet/bottomsheet.dart';
 export 'src/snackbar/snack_route.dart';
+export 'src/state/get_state.dart';
+export 'src/root/root_widget.dart';
 export 'src/routes/observers/route_observer.dart';
 export 'src/routes/transitions_type.dart';
 export 'src/platform/platform.dart';